Added Math Scripts & UI's
Added Multiplier Script and Repetitive Addition Script along with the UI's for both. They both worked successfully when tested. The Multiplier Script is attached to the 1st coin and the Repetitive Addition Script is attached to the 7th coin. When the scripts are activated the player will freeze while the UI is enabled.